cleanup and simplify the gap filling logic in lines reconstruction
